# internal/api/openapi.yaml
openapi: 3.1.0
info:
  title: Agency Gateway API
  version: 0.1.0
  description: |
    REST API for the Agency AI agent operating platform. Deploys agents in
    enforced isolation containers with credential scoping, network mediation,
    read-only constraints, and continuous audit.

    All endpoints except /health require API token authentication via the
    X-Agency-Token header. Streaming endpoints accept `Accept: application/x-ndjson`.
  license:
    name: Proprietary
servers:
  - url: http://localhost:8200/api/v1
    description: Local gateway daemon
security:
  - apiToken: []
...
paths:
...
  /credentials:
    post:
      operationId: createOrUpdateCredential
      summary: Store or update a credential
      description: Stores a credential in the encrypted credential store. Idempotent — same name+value is a no-op, different value updates.
      tags: [Credentials]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, value, kind, scope, protocol]
              properties:
                name: { type: string, description: Credential name (e.g. LC_API_KEY_DETECTION_TUNER) }
                value: { type: string, description: Credential value (encrypted at rest) }
                kind: { type: string, enum: [provider, service, gateway, internal, group] }
                scope: { type: string, description: "platform, agent:{name}, or team:{name}" }
                protocol: { type: string, enum: [api-key, jwt-exchange, github-app, oauth2, bearer] }
                service: { type: string, description: Service definition name }
                group: { type: string, description: Credential group name }
                external_scopes: { type: array, items: { type: string } }
                requires: { type: array, items: { type: string } }
                expires_at: { type: string, format: date-time }
                protocol_config: { type: object, additionalProperties: true }
      responses:
        "200":
          description: Credential stored
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string }
                  name: { type: string }
    get:
      operationId: listCredentials
      summary: List credentials (values redacted)
      tags: [Credentials]
      parameters:
        - name: kind
          in: query
          schema: { type: string }
        - name: scope
          in: query
          schema: { type: string }
        - name: service
          in: query
          schema: { type: string }
        - name: group
          in: query
          schema: { type: string }
        - name: expiring
          in: query
          schema: { type: string }
          description: Duration string (e.g. 7d) — list credentials expiring within
      responses:
        "200":
          description: List of credentials with redacted values
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CredentialEntry"
  /credentials/{name}:
    get:
      operationId: showCredential
      summary: Show credential metadata (value redacted unless ?show_value=true)
      tags: [Credentials]
      parameters:
        - name: name
          in: path
          required: true
          schema: { type: string }
        - name: show_value
          in: query
          schema: { type: boolean, default: false }
          description: Include plaintext value (high-severity audit event)
      responses:
        "200":
          description: Credential details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CredentialEntry"
        "404":
          description: Credential not found
    delete:
      operationId: deleteCredential
      summary: Delete a credential
      tags: [Credentials]
      parameters:
        - name: name
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Credential deleted
  /credentials/{name}/rotate:
    post:
      operationId: rotateCredential
      summary: Rotate credential value (preserves metadata)
      tags: [Credentials]
      parameters:
        - name: name
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [value]
              properties:
                value: { type: string }
      responses:
        "200":
          description: Credential rotated
  /credentials/{name}/test:
    post:
      operationId: testCredential
      summary: Test credential against external API
      tags: [Credentials]
      parameters:
        - name: name
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Test result
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  status: { type: integer }
                  message: { type: string }
                  latency_ms: { type: integer }
  /credentials/groups:
    post:
      operationId: createCredentialGroup
      summary: Create a credential group (shared protocol config)
      tags: [Credentials]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, protocol]
              properties:
                name: { type: string }
                protocol: { type: string }
                token_url: { type: string }
                token_params: { type: object, additionalProperties: { type: string } }
                requires: { type: array, items: { type: string } }
      responses:
        "200":
          description: Group created
  /internal/credentials/resolve:
    get:
      operationId: resolveCredential
      summary: Resolve credential for egress (internal, via Unix socket)
      description: |
        Returns decrypted credential value + protocol metadata. Used by the
        egress proxy via the gateway Unix socket. No Bearer auth required on
        the socket path (socket access = authorization). Auth-gated on HTTP.
      tags: [Internal]
      parameters:
        - name: name
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Resolved credential
          content:
            application/json:
              schema:
                type: object
                properties:
                  name: { type: string }
                  value: { type: string }
                  protocol: { type: string }
                  protocol_config: { type: object }
        "404":
          description: Credential not found

  # ── Agent Rebuild ──────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    apiToken:
      type: apiKey
      in: header
      name: X-Agency-Token
      description: API token generated during `agency init`
    webhookSecret:
      type: apiKey
      in: header
      name: X-Webhook-Secret
      description: Per-webhook secret for inbound webhook authentication
...
  schemas:
...
    CredentialEntry:
      type: object
      properties:
        name: { type: string }
        value: { type: string, description: "[redacted] unless show_value=true" }
        kind: { type: string, enum: [provider, service, gateway, internal, group] }
        scope: { type: string }
        protocol: { type: string }
        service: { type: string }
        group: { type: string }
        external_scopes: { type: array, items: { type: string } }
        requires: { type: array, items: { type: string } }
        expires_at: { type: string, format: date-time }
        protocol_config: { type: object, additionalProperties: true }
        source: { type: string }
        created_at: { type: string, format: date-time }
        rotated_at: { type: string, format: date-time }
...